modified some helpers
