+++ client/src/components/GroupView/GroupTable.js
@@ -1,4 +1,5 @@
 import React, { useCallback, useState, useMemo, useEffect } from 'react';
+import { handleExportData } from '../../utils/exportData';
 import MaterialReactTable from 'material-react-table';
 import {
   Box,
@@ -13,18 +14,15 @@ import {
   Tooltip,
   Typography,
   Select,
-  MenuItem,
-  cellValueMap,
-  InputLabel,
+  MenuItem, InputLabel,
   OutlinedInput,
   Alert
 } from '@mui/material';
 import FileDownloadIcon from '@mui/icons-material/FileDownload';
-import { ExportToCsv } from 'export-to-csv'; //or use your library of choice here
 import { Delete, Edit } from '@mui/icons-material';
 import { FormControl } from '@material-ui/core';
 import Chip from '@mui/material/Chip';
-import { Theme, useTheme } from '@mui/material/styles';
+import { useTheme } from '@mui/material/styles';
 import { FilterDataByProfessor } from '../../utils/FilterDataByValue';
 
 const GroupTable = () => {
@@ -194,61 +192,7 @@ const GroupTable = () => {
     },
     [tableData],
   );
-
-  // For exporting the table data
-  const csvOptions = {
-    filename: 'GroupsFromAcTeams-' + getDate(),
-    fieldSeparator: ',',
-    quoteStrings: '"',
-    decimalSeparator: '.',
-    showLabels: true,
-    useBom: true,
-    useKeysAsHeaders: true,
-  };
-
-  const csvExporter = new ExportToCsv(csvOptions);
-
-  const handleExportData = () => {
-    // clean up and organize data to be exported
-    const keyToRemove = "_id"
-    const updatedJsonList = tableData.map(jsonObj => {
-      let updatedJsonObject = jsonObj
-      // remove the _id as that should not be in the json
-      if (keyToRemove in jsonObj) {
-        const { [keyToRemove]: deletedKey, ...rest } = jsonObj // use destructuring to remove the key
-        updatedJsonObject = rest // return the updated JSON object without the deleted key
-      }
-
-      // sort the keys as they appear in the columns
-      const orderedKeys = columns.map(key => key.accessorKey)
-      updatedJsonObject = Object.keys(updatedJsonObject)
-        .sort((a, b) => orderedKeys.indexOf(a) - orderedKeys.indexOf(b)) // sort keys in the order of the updated keys
-        .reduce((acc, key) => ({ ...acc, [key]: updatedJsonObject[key] }), {}) // create a new object with sorted keys
-
-      // replace the accessor key by the header
-      for (let i = 0; i < columns.length; i++) {
-        const { accessorKey, header } = columns[i]
-        if (accessorKey in updatedJsonObject) {
-          const { [accessorKey]: renamedKey, ...rest } = updatedJsonObject // use destructuring to rename the key
-          updatedJsonObject = { ...rest, [header]: renamedKey } // update the JSON object with the renamed key
-        }
-      }
-
-      return updatedJsonObject // return the original JSON object if the key is not found
-    })
-
-    csvExporter.generateCsv(updatedJsonList);
-  };
-
-  function getDate() {
-    const today = new Date();
-    const year = today.getFullYear();
-    const month = String(today.getMonth() + 1).padStart(2, '0');
-    const day = String(today.getDate()).padStart(2, '0');
-    const formattedDate = `${year}-${month}-${day}`;
-    return formattedDate
-  }
-
+ 
   return (
     <Box sx={{ p: 2 }}>
       <Typography variant="h2" align="center" fontWeight="fontWeightBold" sx={{ marginBottom: '0.5rem' }}>Groups</Typography>
@@ -302,7 +246,7 @@ const GroupTable = () => {
             </Button>
             <Button
               color="primary"
-              onClick={handleExportData}
+              onClick={handleExportData(tableData,columns)}
               startIcon={<FileDownloadIcon />}
               variant="contained"
             >
